Rename workspace interfaces files
:w
