+++ backend/src/data/repositories/workspaceRepository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository, Repository } from 'typeorm';
 import { Workspace } from '../entities/Workspace';
-import { ICreateWorkspace } from '../../common/models/workspace/createWorkspace';
+import { ICreateWorkspace } from '../../common/models/workspace/ICreateWorkspace';
 
 @EntityRepository(Workspace)
 class WorkspaceRepository extends Repository<Workspace> {
