+++ backend/src/services/workspaceService.ts
@@ -1,8 +1,8 @@
 import { getCustomRepository } from 'typeorm';
 
 import WorkspaceRepository from '../data/repositories/workspaceRepository';
-import { ICreateWorkspace } from '../common/models/workspace/createWorkspace';
-import { IWorkspaceResponse } from '../common/models/workspace/workspaceResponse';
+import { ICreateWorkspace } from '../common/models/workspace/ICreateWorkspace';
+import { IWorkspaceResponse } from '../common/models/workspace/IWorkspaceResponse';
 import { fromCreatedWorkspaceToClient } from '../common/mappers/workspace';
 
 export const createWorkspace = async (data: ICreateWorkspace): Promise<IWorkspaceResponse> => {
